partrt: New test case, bug fix, skipped newline in calls to info(), debug() and fail()
cpumask_list() has now a new test case to test for a bug that happens when
called with a cpu containing the last available cpu. The test case was
verified to detect the bug, the bug was then fixed and the test case was
verified to pass after this.

Newline should no longer be given in calls to info(), debug() and fail(). This
will make it behave the same as for calls to tracef(), which does not expect
newlines.

Signed-off-by: Mats Liljegren <mats.liljegren@enea.com>