partrt: Moved script, added cmake support, made it ash compatible
partrt has been copied to partrt sub-directory from install/bin. The install/bin
version is now obsolete and scheduled to be removed in the near future.

The script has been changed to be ash compatible, dash to be specific. This is
to enable it to be used by busybox sh.

The script now uses bitcalc.

The partrt_env file has changed syntax: The name and value pair is separated by
blank rather than '='.

Signed-off-by: Mats Liljegren <mats.liljegren@enea.com>