partrt: Made script ash-compatible, use bitcalc, move man page to share
This also increments the version from 1.1 to 1.2.
This also means that list2mask is no longer needed. Since this was the last
file in the install directory, this directory is now removed.

The man page was moved from man/man1 -> share/man/man1, since this seems to be
the standard location.

The README file was deleted in the rt-tools root, since the README.md using
markdown format has replaced it with an updated content.

Signed-off-by: Mats Liljegren <mats.liljegren@enea.com>
Signed-off-by: Andreas Sundström <andreas.sundstrom@enea.com>