partrt: Use bitmask array when traversing bitmasks
The [non]isolated_cpu_list variables are not for iterations, but are being used
for that anyway. This will not iterate the bits. This currently affect
hotplug and the newly introduced cpufreq governor selection.

Bitcalc has a newly introduced format callaed "array", which is designed for
allowing scripts to iterate over set bits. Use that for iterations.

Signed-off-by: Mats Liljegren <mats.liljegren@enea.com>